Added game logic and restructuring

+++ src/game/uiPreset.cpp
@@ -2,6 +2,7 @@
 #include <cmath>
 #include <utility>
 #include <algorithm>
+#include <iterator>
 #include <cstddef>
 
 #include <glm/glm.hpp>
@@ -31,7 +32,7 @@ void InteractableObjectBackground::configureShaders() const
     unsigned int modelLoc = glGetUniformLocation(shader->getID(), "model");
 
     glUniform3fv(colorLoc, 1, glm::value_ptr(m_outlineColor));
-    
+
     static GLFWController& glfwControllerInstance {GLFWController::getInstance()};
     float windowWidth = glfwControllerInstance.getWidth();
     float windowHeight = glfwControllerInstance.getHeight();
@@ -74,24 +75,12 @@ void SettingUIElement::trigger()
     UIElement::trigger();
 }
 
-UIElement& UIElement::operator=(const UIElement& other) noexcept
-{
-    if(&other != this)
-    {
-        if(other.m_backgroundObject)
-            m_backgroundObject = std::make_unique<Object2D>(*other.m_backgroundObject);
-        else m_backgroundObject = nullptr;
-        m_textData = other.m_textData;
-        m_callback = other.m_callback;
-    }
-    return *this;
-}
 std::string SettingUIElement::getText()
 {
     return *m_isEnabled ? m_enabledText : m_textData.text;
 }
 
-void UIPreset::initialize()
+void UIPreset::initialize(std::vector<std::unique_ptr<UIElement>>&& elements)
 {
     m_text = gltCreateText();
 
@@ -109,45 +98,33 @@ void UIPreset::initialize()
         -1.f, -1.f, 0.f
     };
     static Mesh uiElementMesh {meshtools::generateMesh(vertices, std::ssize(vertices))};
-    auto elementsSize {m_elements.size()};
-    bool hasInteractiveElements {};
+    auto elementsSize {elements.size()};
     for(std::size_t i {}; i < elementsSize; ++i)
     {
         //interactive elements
-        if(m_elements[i]->m_callback)
+        if(elements[i]->m_callback)
         {
-            m_elements[i]->m_backgroundObject =
+            elements[i]->m_backgroundObject =
                 std::make_unique<InteractableObjectBackground>(uiElementMesh, uiElementShader,
-                glm::vec3(m_elements[i]->m_textData.backgroundColor.x, m_elements[i]->m_textData.backgroundColor.y, m_elements[i]->m_textData.backgroundColor.z), m_highlightColor);
-
-            if(!hasInteractiveElements)
-                hasInteractiveElements = true;
+                glm::vec3(elements[i]->m_textData.backgroundColor.x, elements[i]->m_textData.backgroundColor.y, elements[i]->m_textData.backgroundColor.z), m_highlightColor);
         }
         //noninteractive elements
         else
         {
-            m_elements[i]->m_backgroundObject =
+            elements[i]->m_backgroundObject =
                 std::make_unique<Object2D>(uiElementMesh, uiElementShader, 
-                glm::vec3(m_elements[i]->m_textData.backgroundColor.x, m_elements[i]->m_textData.backgroundColor.y, m_elements[i]->m_textData.backgroundColor.z));
+                glm::vec3(elements[i]->m_textData.backgroundColor.x, elements[i]->m_textData.backgroundColor.y, elements[i]->m_textData.backgroundColor.z));
         }
     }
-    if(!hasInteractiveElements) return;
-    
-    //save the first interactable argument's position because the first argument is meant to be focused by default.
-    auto defaultFocusElementPos {
-        std::find_if(m_elements.begin(), m_elements.end(), [](const auto& element)
-        {
-            return element->m_callback;
-        })->get()->m_textData.position};
 
     //sort the elements to be iterable correctly for keyboard input
-    std::sort(m_elements.begin(), m_elements.end(), [](const auto& a, const auto& b) -> bool
+    std::sort(elements.begin(), elements.end(), [](const auto& a, const auto& b) -> bool
     {
         float yDifference {a->m_textData.position.y - b->m_textData.position.y};
         if(yDifference > SAME_ROW_EPSILON)
-            return true;
-        else if(yDifference < -SAME_ROW_EPSILON)
             return false;
+        else if(yDifference < -SAME_ROW_EPSILON)
+            return true;
         else return (a->m_textData.position.x - b->m_textData.position.x) < .0f;
     });
 
@@ -157,92 +134,76 @@ void UIPreset::initialize()
         return std::fabs(element1.m_textData.position.y - element2.m_textData.position.y) < SAME_ROW_EPSILON;
     };
 
-    bool foundDefaultFocusElement {};
-    for(std::size_t i {0}; m_elements[i]->m_callback;)
+    for(std::size_t i {0}; i < elementsSize;)
     {
-        bool lastRange {true};
         std::size_t rangeLength {1};
-        while((i + rangeLength) != elementsSize && m_elements[i + rangeLength]->m_callback)
+        while((i + rangeLength) != elementsSize && elements[i + rangeLength]->m_callback)
         {
-            if(!isSameRow(*m_elements[i + rangeLength - 1], *m_elements[i + rangeLength]))
+            if(!isSameRow(*elements[i + rangeLength - 1], *elements[i + rangeLength]))
             {
-                lastRange = false;
                 break;
             }
             ++rangeLength;
         }
 
-        m_interactableElements.push_back(std::span<std::unique_ptr<UIElement>>(m_elements.begin() + i, rangeLength));
-
-        if(!foundDefaultFocusElement)
-        {
-            auto it = std::find_if(m_interactableElements.back().begin(), 
-                m_interactableElements.back().end(), [&](const auto& element)
-                {
-                    return element->m_textData.position == defaultFocusElementPos;
-                });
-                
-            if(it != m_interactableElements.back().end())
-            {
-                foundDefaultFocusElement = true;
-                m_defaultFocusIndices = ElementIndices(i, 
-                    std::distance(m_interactableElements.back().begin(), it));
-            }
-        }
-
-        if(lastRange)
-            break;
+        m_sortedElements.emplace_back(std::make_move_iterator(elements.begin() + i), 
+            std::make_move_iterator(elements.begin() + i + rangeLength));
 
         i += rangeLength;
     }
 }
 void UIPreset::updateBackgroundsUniforms(int windowWidth, int windowHeight)
 {
-    for(std::size_t i {}; i < m_elements.size(); ++i)
+    for(auto& row : m_sortedElements)
     {
-        float sizeMultiplier {(windowHeight < windowWidth ? windowHeight : windowWidth) * TEXT_SIZE_MULTIPLIER * m_elements[i]->m_textData.backgroundScale};
-        glm::mat4 backgroundModel(1.f);
+        for(auto& element : row)
+        {
+            float sizeMultiplier {(windowHeight < windowWidth ? windowHeight : windowWidth) * TEXT_SIZE_MULTIPLIER * element->m_textData.backgroundScale};
+            glm::mat4 backgroundModel(1.f);
 
-        gltSetText(m_text, m_elements[i]->m_textData.text.c_str());//this isn't using getText() because background boxes are meant to have constant sizes
-        
-        auto textWidth {gltGetTextWidth(m_text, m_elements[i]->m_textData.scale) / (float)windowWidth};
-        auto textHeight {gltGetTextHeight(m_text, m_elements[i]->m_textData.scale) / (float)windowHeight};
+            gltSetText(m_text, element->m_textData.text.c_str());//this isn't using getText() because background boxes are meant to have constant sizes
+            
+            auto textWidth {gltGetTextWidth(m_text, element->m_textData.scale) / (float)windowWidth};
+            auto textHeight {gltGetTextHeight(m_text, element->m_textData.scale) / (float)windowHeight};
 
-        backgroundModel = glm::translate(backgroundModel, glm::vec3(m_elements[i]->m_textData.position.x, m_elements[i]->m_textData.position.y, 0.f));
-        backgroundModel = glm::scale(backgroundModel, glm::vec3(textWidth * sizeMultiplier, textHeight * sizeMultiplier, 0.f));
-        m_elements[i]->m_backgroundObject->setModel(std::move(backgroundModel));
+            backgroundModel = glm::translate(backgroundModel, glm::vec3(element->m_textData.position.x, element->m_textData.position.y, 0.f));
+            backgroundModel = glm::scale(backgroundModel, glm::vec3(textWidth * sizeMultiplier, textHeight * sizeMultiplier, 0.f));
+            element->m_backgroundObject->setModel(std::move(backgroundModel));
+        }
     }
 }
 
 void UIPreset::setFocusedElement(ElementIndices elementIndices)
 {
-    dynamic_cast<InteractableObjectBackground*>(m_interactableElements[m_focusIndicies.first][m_focusIndicies.second]->m_backgroundObject.get())->setUseOutline(false);
+    auto oldBackground {dynamic_cast<InteractableObjectBackground*>(m_sortedElements[m_focusIndicies.first][m_focusIndicies.second]->m_backgroundObject.get())};
+    if(oldBackground) oldBackground->setUseOutline(false);
     m_focusIndicies = elementIndices;
-    dynamic_cast<InteractableObjectBackground*>(m_interactableElements[m_focusIndicies.first][m_focusIndicies.second]->m_backgroundObject.get())->setUseOutline(true);
+    dynamic_cast<InteractableObjectBackground*>(m_sortedElements[m_focusIndicies.first][m_focusIndicies.second]->m_backgroundObject.get())->setUseOutline(true);
 }
 
 void UIPreset::moveFocusedElement(FocusMoveDirections focusMoveDirection)
 {
-    if(m_interactableElements.size() == 1 && m_interactableElements[0].size() == 1) return;
+    if(m_interactableElementsCount < 2) return;
+
     ElementIndices newElementIndices {};
+
     auto moveUp = [&]()
     {
-        if(m_focusIndicies.first == 0)
-        {
-            newElementIndices.first = m_interactableElements.size() - 1;
-        }
-        else newElementIndices.first = m_focusIndicies.first - 1;
-        newElementIndices.second = m_interactableElements[newElementIndices.first].size() - 1;
-    };
-
-    auto moveDown = [&]()
-    {
-        if(m_focusIndicies.first == (m_interactableElements.size() - 1))
+        if(m_focusIndicies.first == m_sortedElements.size() - 1)
         {
             newElementIndices.first = 0;
         }
         else newElementIndices.first = m_focusIndicies.first + 1;
         newElementIndices.second = 0;
+    };
+    auto moveDown = [&]()
+    {
+        if(m_focusIndicies.first == 0)
+        {
+            newElementIndices.first = m_sortedElements.size() - 1;
+        }
+        else newElementIndices.first = m_focusIndicies.first - 1;
+        newElementIndices.second = m_sortedElements[newElementIndices.first].size() - 1;
     };
 
     switch (focusMoveDirection)
@@ -254,7 +215,7 @@ void UIPreset::moveFocusedElement(FocusMoveDirections focusMoveDirection)
         moveDown();
         break;
     case FocusMoveDirections::right:
-        if(m_focusIndicies.second == (m_interactableElements[m_focusIndicies.first].size() - 1))
+        if(m_focusIndicies.second == (m_sortedElements[m_focusIndicies.first].size() - 1))
         {
             moveDown();
         }
@@ -269,7 +230,7 @@ void UIPreset::moveFocusedElement(FocusMoveDirections focusMoveDirection)
         {
             moveUp();
         }
-        else 
+        else
         {
             newElementIndices.first = m_focusIndicies.first;
             newElementIndices.second = m_focusIndicies.second - 1;
@@ -277,17 +238,33 @@ void UIPreset::moveFocusedElement(FocusMoveDirections focusMoveDirection)
         break;
     }
 
+    //if the next element is not interactable, recursion is used to move again
+    if(!m_sortedElements[newElementIndices.first][newElementIndices.second]->m_callback)
+    {
+        auto oldBackground {dynamic_cast<InteractableObjectBackground*>(m_sortedElements[m_focusIndicies.first][m_focusIndicies.second]->m_backgroundObject.get())};
+        if(oldBackground) oldBackground->setUseOutline(false);
+        m_focusIndicies = newElementIndices;
+        moveFocusedElement(focusMoveDirection);
+
+        return;
+    }
+
     setFocusedElement(newElementIndices);
 }
 void UIPreset::initializeGLT()
 {
     GLFWController& glfwControllerInstance {GLFWController::getInstance()};
-    if (!gltInit())
+    static bool initialized {};
+    if(!initialized)
     {
-        std::cerr << "Failed to initialize glText\n";
-        glfwControllerInstance.terminate();
+        if (!gltInit())
+        {
+            std::cerr << "Failed to initialize glText\n";
+            glfwControllerInstance.terminate();
+        }
+        gltViewport(glfwControllerInstance.getWidth(), glfwControllerInstance.getHeight());
+        initialized = true;
     }
-    gltViewport(glfwControllerInstance.getWidth(), glfwControllerInstance.getHeight());
 }
 UIPreset::~UIPreset()
 {
@@ -300,16 +277,30 @@ void UIPreset::enable()
     static RenderEngine& renderEngineInstance {RenderEngine::getInstance()};
 
     updateBackgroundsUniforms(glfwControllerInstance.getWidth(), glfwControllerInstance.getHeight());
-    setFocusedElement(m_defaultFocusIndices);
-    for(std::size_t i {0}; i < m_elements.size(); ++i)
-        renderEngineInstance.addObject(m_elements[i]->m_backgroundObject.get());
+
+    m_interactableElementsCount = 0;
+    //send the objects to render engine and make the first interactable element focused
+    for(std::size_t i {}; i < m_sortedElements.size(); ++i)
+    {
+        for(std::size_t j {}; j < m_sortedElements[i].size(); ++j)
+        {
+            if(m_sortedElements[i][j]->m_callback) 
+            {   
+                ++m_interactableElementsCount;
+
+                setFocusedElement(std::make_pair(i, j));
+            }
+            renderEngineInstance.addObject(m_sortedElements[i][j]->m_backgroundObject.get());
+        }
+    }
 }
 
 void UIPreset::disable()
 {
     static RenderEngine& renderEngineInstance {RenderEngine::getInstance()};
-    for(std::size_t i {0}; i < m_elements.size(); ++i)
-        renderEngineInstance.removeObject(m_elements[i]->m_backgroundObject.get());
+    for(auto& row : m_sortedElements)
+        for(auto& element : row)
+            renderEngineInstance.removeObject(element->m_backgroundObject.get());
 }
 
 void UIPreset::update()
@@ -319,16 +310,21 @@ void UIPreset::update()
     int width {glfwControllerInstance.getWidth()}, height {glfwControllerInstance.getHeight()};
     float sizeMultiplier {(height < width ? height : width) * TEXT_SIZE_MULTIPLIER};
 
-    for(std::size_t i {}; i < m_elements.size(); ++i)
+    for(auto& row : m_sortedElements)
     {
-        gltColor(m_elements[i]->m_textData.textColor.x, 
-            m_elements[i]->m_textData.textColor.y, m_elements[i]->m_textData.textColor.z, 1.f);    
-        gltSetText(m_text, m_elements[i]->getText().c_str());
-        gltDrawText2DAligned(m_text,
-            (GLfloat)(width / 2.f) + (m_elements[i]->m_textData.position.x / 2 * width),
-            (GLfloat)(height / 2.f) + (m_elements[i]->m_textData.position.y / 2 * -height),
-            m_elements[i]->m_textData.scale * sizeMultiplier,
-            GLT_CENTER, GLT_CENTER);
+        for(auto& element : row)
+        {
+            if(element->m_textData.text == "") continue;
+
+            gltColor(element->m_textData.textColor.x, 
+                element->m_textData.textColor.y, element->m_textData.textColor.z, 1.f);    
+            gltSetText(m_text, element->getText().c_str());
+            gltDrawText2DAligned(m_text,
+                (GLfloat)(width / 2.f) + (element->m_textData.position.x / 2 * width),
+                (GLfloat)(height / 2.f) + (element->m_textData.position.y / 2 * -height),
+                element->m_textData.scale * sizeMultiplier,
+                GLT_CENTER, GLT_CENTER);
+        }
     }
 
     gltEndDraw();
@@ -340,7 +336,8 @@ void UIPreset::processInput(int key)
     {
     case GLFW_KEY_ENTER:
     case GLFW_KEY_SPACE:
-        m_interactableElements[m_focusIndicies.first][m_focusIndicies.second]->trigger();
+        if(m_interactableElementsCount)
+            m_sortedElements[m_focusIndicies.first][m_focusIndicies.second]->trigger();
         break;
 
     case GLFW_KEY_UP:
